min/max calls replaced with std::min/std::max

+++ include/nxcpapi.h
@@ -140,7 +140,7 @@ public:
    TCHAR *getFieldAsString(UINT32 fieldId, TCHAR *buffer = NULL, size_t bufferSize = 0) const;
 	char *getFieldAsMBString(UINT32 fieldId, char *buffer = NULL, size_t bufferSize = 0) const;
 	char *getFieldAsUtf8String(UINT32 fieldId, char *buffer = NULL, size_t bufferSize = 0) const;
-   UINT32 getFieldAsBinary(UINT32 fieldId, BYTE *buffer, size_t bufferSize) const;
+   size_t getFieldAsBinary(UINT32 fieldId, BYTE *buffer, size_t bufferSize) const;
    InetAddress getFieldAsInetAddress(UINT32 fieldId) const;
    MacAddress getFieldAsMacAddress(UINT32 fieldId) const;
    uuid getFieldAsGUID(UINT32 fieldId) const;
